Changed `Nonogram` to use `Cell<Fill>` instead of `Fill`

+++ crates/puzzled_core/src/puzzle/style/color.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-pub type ColorId = usize;
+pub type ColorId = u32;
 
 type ColorValue = u8;
